Решение задачи 3.1.2. Spring Boot, Security

+++ src/main/java/ru/javamentor/springsecurity/service/UserService.java
@@ -1,4 +1,5 @@
 package ru.javamentor.springsecurity.service;
+import ru.javamentor.springsecurity.model.Role;
 import ru.javamentor.springsecurity.model.User;
 
 import java.util.List;
@@ -8,6 +9,6 @@ public interface UserService {
     void editUser(User user);
     void deleteUser(User user);
     List<User> getListUsers();
-    User getUser (Integer id);
+    User getUser (Long id);
     User findByLogin (String login);
 }
